requirements.txt
asgiref==3.2.5
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.5
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.2.13
django-crispy-forms==1.7.2
django-parler==2.0.1
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.2
djangorestframework==3.11.0
html5lib==1.0.1
<<<<<<< HEAD
httplib2==0.18.0
=======
httplib2==0.11.3
jsmin==2.2.2
Markdown==3.2.1
>>>>>>> Develop
pdfrw==0.4
Pillow==7.0.0
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-slugify==4.0.1
pytz==2018.4
reportlab==3.5.2
six==1.11.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==0.6.1
virtualenv==16.0.0
webencodings==0.5.1
wfastcgi==3.0.0
xhtml2pdf==0.2.2
xlwt==1.3.0
